+++ app/assets/javascripts/folio/console/base.js
@@ -1,11 +1,12 @@
 //= require jquery
-//= require jquery_ujs
+//= require rails-ujs
 //= require folio-bootstrap/dist/js/bootstrap.bundle
-//= require moment/min/moment.min
+//= require moment/moment
 //= require moment/locale/cs
 //= require eonasdan-bootstrap-datetimepicker/build/js/bootstrap-datetimepicker.min
 //= require slideout.js/dist/slideout
 //= require jquery-debounce/jquery.debounce
+//= require multiselect/js/jquery.multi-select
 
 //= require jquery-ui/jquery-ui
 //= require ilikenwf-nested-sortable/jquery.mjs.nestedSortable
@@ -48,6 +49,7 @@
 //= require folio/console/form/errors/errors
 //= require folio/console/modules/_layout
 //= require folio/console/modules/_dirty-forms
+//= require folio/console/modules/_multiselect
 //= require folio/console/layout/sidebar/search/search
 
 //= require folio/console/simple_form_inputs/_date_time_input