+++ app/cells/folio/console/index/images_cell.rb
@@ -10,7 +10,7 @@ class Folio::Console::Index::ImagesCell < Folio::ConsoleCell
 
   def show
     if options[:cover]
-      render if model.cover.present?
+      render
     else
       render if model.images.present?
     end
@@ -20,5 +20,9 @@ class Folio::Console::Index::ImagesCell < Folio::ConsoleCell
     image_from(placement,
                Folio::FileSerializer::ADMIN_THUMBNAIL_SIZE,
                lightbox(placement).merge(class: 'f-c-index-images__img'))
+  end
+
+  def broken_src
+    'data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyNCIgaGVpZ2h0PSIyNCIgdmlld0JveD0iMCAwIDI0IDI0Ij48cGF0aCBmaWxsPSJub25lIiBkPSJNMCAwaDI0djI0SDB6bTAgMGgyNHYyNEgwem0yMSAxOWMwIDEuMS0uOSAyLTIgMkg1Yy0xLjEgMC0yLS45LTItMlY1YzAtMS4xLjktMiAyLTJoMTRjMS4xIDAgMiAuOSAyIDIiLz48cGF0aCBmaWxsPSJub25lIiBkPSJNMCAwaDI0djI0SDB6Ii8+PHBhdGggZD0iTTIxIDV2Ni41OWwtMy0zLjAxLTQgNC4wMS00LTQtNCA0LTMtMy4wMVY1YzAtMS4xLjktMiAyLTJoMTRjMS4xIDAgMiAuOSAyIDJ6bS0zIDYuNDJsMyAzLjAxVjE5YzAgMS4xLS45IDItMiAySDVjLTEuMSAwLTItLjktMi0ydi02LjU4bDMgMi45OSA0LTQgNCA0IDQtMy45OXoiLz48L3N2Zz4='
   end
 end