+++ app/controllers/folio/console/menus_controller.rb
@@ -44,4 +44,12 @@ class Folio::Console::MenusController < Folio::Console::BaseController
     def sti_menu_items(params)
       sti_hack(params, :menu_items_attributes, :target)
     end
+
+    def folio_console_collection_includes
+      [ :menu_items ]
+    end
+
+    def folio_console_record_includes
+      [ :menu_items ]
+    end
 end