+++ app/helpers/folio/cstypo_helper.rb
@@ -6,7 +6,10 @@ module Folio
 
     def cstypo(string)
       if I18n.locale == :cs
-        string.gsub(CSTYPO_REGEXP, '\1&nbsp;').html_safe
+        string.gsub(CSTYPO_REGEXP, '\1&nbsp;')
+              .gsub(/(\d+)\s+(\d+)/, '\1&nbsp;\2')
+              .gsub(/(\d+)\s+(Kč)/, '\1&nbsp;\2')
+              .html_safe
       else
         string
       end