+++ app/models/folio/file_placement/cover.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Folio::FilePlacement::Cover < Folio::FilePlacement::Base
+  include Folio::Sitemap::FilePlacement::Image
+
   folio_image_placement :cover_placement
 end