+++ app/models/folio/file_placement/image.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Folio::FilePlacement::Image < Folio::FilePlacement::Base
+  include Folio::Sitemap::FilePlacement::Image
+
   folio_image_placement
 end