+++ app/models/folio/image.rb
@@ -3,8 +3,39 @@
 class Folio::Image < Folio::File
   include Folio::DragonflyFormatValidation
   include Folio::Thumbnails
+  include Folio::Sitemap::Image
 
   validate_file_format
+
+  dragonfly_accessor :file do
+    after_assign :sanitize_filename
+    after_assign { |image| image.metadata }
+  end
+
+  # Get from metadata
+  def title
+    metadata_compose(['Headline', 'Title'])
+  end
+
+  def caption
+    metadata_compose(['Caption', 'Description', 'Abstract'])
+  end
+
+  def keywords
+    metadata_compose(['Keywords'])
+  end
+
+  def geo_location
+    # Geographic location, e.g.: Limerick, Ireland
+    metadata_compose(['LocationName', 'SubLocation', 'City', 'ProvinceState', 'CountryName'])
+  end
+
+  private
+    def metadata_compose(tags)
+      string_arr = tags.collect { |tag| file_metadata.try('[]', tag) }.compact.uniq
+      return nil if string_arr.size == 0
+      string_arr.join(', ')
+    end
 end
 
 # == Schema Information