+++ lib/folio/engine.rb
@@ -46,7 +46,7 @@ module Folio
     config.folio_console_sidebar_runner_up_link_class_names = []
 
     initializer :append_migrations do |app|
-      unless app.root.to_s.match? root.to_s
+      unless app.root.to_s.include? root.to_s
         config.paths['db/migrate'].expanded.each do |expanded_path|
           app.config.paths['db/migrate'] << expanded_path
         end