+++ test/dummy/db/seeds.rb
@@ -22,6 +22,13 @@ def unsplash_pic(square = false)
   image
 end
 
+def file_pic(file_instance)
+  image = Folio::Image.new
+  image.file = file_instance
+  image.save!
+  image
+end
+
 2.times { unsplash_pic }
 
 Folio::Site.create!(title: 'Sinfin.digital',
@@ -37,15 +44,31 @@ Folio::Site.create!(title: 'Sinfin.digital',
                     })
 
 about = Folio::Page.create!(title: 'O nás',
-                            published: true)
+                            published: true,
+                            published_at: 1.month.ago)
 about.cover = unsplash_pic
 3.times { about.images << unsplash_pic }
+about.image_placements.each { |ip|
+  name = Faker::Name.name
+  ip.update_attributes!(alt: name, title: "Portrait of #{name}")
+}
+
+
+night_sky = Folio::Page.create!(title: 'Noční obloha', published: true, published_at: 1.month.ago, locale: :cs)
+night_photo = File.new(Rails.root.join('..', 'fixtures', 'folio', 'photos', 'night.jpg'))
+night_sky.cover = file_pic(night_photo)
+1.times { night_sky.images << file_pic(night_photo) }
+
+# TODO: Atoms
 
 reference = Folio::Page.create!(title: 'Reference',
                                 published: true,
                                 published_at: 1.day.ago)
-Folio::Page.create!(title: 'Smart Cities', published: true)
-Folio::Page.create!(title: 'Vyvolej.to', published: true)
+Folio::Page.create!(title: 'Smart Cities', published: true, published_at: 1.month.ago)
+vyvolejto = Folio::Page.create!(title: 'Vyvolej.to', published: true, published_at: 1.month.ago)
+iptc_test = File.new(Rails.root.join('..', 'fixtures', 'folio', 'photos', 'exif-samples', 'jpg', 'tests', '46_UnicodeEncodeError.jpg'))
+vyvolejto.cover = file_pic(iptc_test)
+
 Folio::Page.create!(title: 'Hidden', published: false)
 Folio::Page.create!(title: 'DAM', published: true)